txDevITelexSrv: Fix nasty bug re. welcome banner
In rare cases, the i-Telex server class could "lock up" in a way that
nothing would be received from any caller, ever again. The story goes
as follows:

As server, the read method -- by which data to be printed is passed to
main loop -- is modified to only return printable data if

- the welcome banner has been enqueued and
- it has been completely printed.

The enqueuing is checked by receipt of ESC-WELCOME and advances state
from CON_TP_RUN to CON_FULL. If it's printed is checked using the
i-Telex acknowledge counter we send to remote as feedback for our print
buffer: Implemented like in i-Telex hardware, it starts at -24 and
reaches 0 as the last character is sent to the teleprinter, the banner
always being 24 characters long.

This whole mechanism is there to ensure that the banner is printed
unmodified. Some i-Telex services for instance immediately send a WRU,
triggering the teleprinter's responder during banner printing and
causing letter salad. This is alleviated by withholding received data
until the banner has in fact been printed, which is determined as stated
above.

The bug is triggered as follows: A remote party connects, sends data and
immediately disconnects -- before the banner could be printed. This
leaves _acknowledge_counter < 0 and causes printable data to be withheld
forever in _rx_buffer, even accumulating over future incoming
connections. Even as _acknowledge_counter is reset on the next
connection, the calculation goes wrong because the receive buffer is
still full, never being able to reach 0 again.

As solution, the withholding condition was modified to only check
_acknowledge_counter while connected. As it was easier to write and
comprehend, the logic was ingverted, making the withhold case "else" and
the send-all case the "if".

Before, after a disconnect, ESC-Z was enqueued behind the received data,
but due to _acknowledge_counter being less than 0, immediately returned,
leaving the buffer full. On the next connection, the value of
_acknowledge_counter would always be too small, incorporating data from
the last connection(s).

Now, after a disconnect, self._connected == ST.DISCON and the if
triggers, returning *all* data and thus emptying the buffer, allowing
_acknowledge_counter to be properly evaluated.