txDevITelex*: After disconnect, pause until tp buffer empty
Correct acknowledge counter calculation is vital for proper buffer
management by the remote peer. As it is calculated from the received
bytes in this connection minus all intermediate buffer contents, it is
important that a connection starts with empty buffers.

During normal operation, this was mostly being achieved, the frequent
exception being multiple consecutive connections by automated services.
Also, this may happen with manual connections in rare cases.

When an automated service connected multiple times consecutively, the
teleprinter still had a nonempty print buffer when the connection was
established, leading to a negative acknowledge counter.

Fix this issue by introducing a new state before ST.DISCON, which is
only entered when disconnect happens with nonempty teleprinter buffer.
In this state, we only wait for ESC-~0 in write method. If it arrives,
we know the printer has finished and reset our state to accept new
connections.

The new state is only effective in txDevITelexSrv and triggers "occ" if
a new client tries to connect while we're still waiting for the
teleprinter to finish. In txDevITelexClient, we don't need to act since
a new connection can only be made by going through dialling, which must
be initiated from the teleprinter.

Also, implement minor logging improvements and move the setting of
self._print_buf_len into write method.